+++ app/Http/Controllers/PackageController.php
@@ -5,7 +5,9 @@ namespace App\Http\Controllers;
 use App\Http\Controllers\Controller;
 use App\Package;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Hash;
 
 class PackageController extends Controller
 {
@@ -172,16 +174,22 @@ class PackageController extends Controller
      * Remove the specified resource from storage.
      *
      * @param \App\Package $package
+     * @param Request $request
      * @return void
      */
-    public function destroy($package)
+    public function destroy($package, Request $request)
     {
-        $pack = Package::where('id',$package)->first();
-        if ($pack['image'] != null) {
-            $image_path = public_path() . '/storage/' . $pack['image'];
-            unlink($image_path);
+        $password = $request->input('password');
+        $user_password = Auth::user()->getAuthPassword();
+        if(Hash::check($password, $user_password)) {
+            $pack = Package::where('id', $package)->first();
+            if ($pack['image'] != null) {
+                $image_path = public_path() . '/storage/' . $pack['image'];
+                unlink($image_path);
+            }
+            Package::destroy($package);
+            return redirect('package')->with('message', 'Delete Successfully');
         }
-        Package::destroy($package);
-        return redirect('package');
+        return redirect('package')->with('delete', 'Password Not Valid');
     }
 }
